LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-outside/values/shape-margin-005-expected.txt


PASS shape-margin can be assigned 'inherit' value
PASS shape-margin is not inherited and defaults to 0px
PASS shape-margin is not inherited
